Refactor login controller and separated in different file
