+++ src/routers/jobSeeker.ts
@@ -3,10 +3,10 @@ import tokenVerify from '../middlewares/tokenVerify'
 
 import {
   getJobSeeker,
-  jobSeekerLocalLogin,
   createJobSeeker,
   updateJobSeeker,
 } from '../controllers/jobSeeker'
+import { jobSeekerLocalLogin as user } from '../controllers/login'
 import { match } from '../controllers/match'
 
 const router = express.Router()
@@ -14,7 +14,7 @@ const router = express.Router()
 router.get('/', tokenVerify, getJobSeeker)
 router.post('/', createJobSeeker)
 router.get('/match', tokenVerify, match)
-router.post('/login/local', jobSeekerLocalLogin)
+router.post('/login/local', user)
 router.patch('/', tokenVerify, updateJobSeeker)
 
 export default router
