+++ src/routers/login.ts
@@ -1,0 +1,8 @@
+import express from 'express'
+import { jobSeekerLocalLogin as user } from '../controllers/login'
+
+const router = express.Router()
+
+router.post('/login/local', user)
+
+export default router
